pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-agent-orchestrator"
version = "0.1.0"
description = "Multi-agent voice AI orchestration framework for home services"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "livekit-agents>=1.0.0",
    "livekit-plugins-deepgram>=1.0.0",
    "livekit-plugins-openai>=1.0.0",
    "livekit-plugins-cartesia>=1.0.0",
    "livekit-plugins-silero>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
